requirements.txt
py-moneyed>=0.8.0
pytest>=4.5.0
tablib>=0.13.0
